Update UsdValue component
