+++ src/utils/operations/getUsdValue.tsx
@@ -1,12 +1,10 @@
-export const getUsdValue = ({
-  amount,
-  usd,
-  decimals = 2
-}: {
+export interface UsdValueType {
   amount: string;
   usd: number;
   decimals?: number;
-}) => {
+}
+
+export const getUsdValue = ({ amount, usd, decimals = 2 }: UsdValueType) => {
   const sum = (parseFloat(amount) * usd).toFixed(decimals);
   const formattedValue = parseFloat(sum).toLocaleString('en', {
     maximumFractionDigits: decimals,
